requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.0.350
langchain-community==0.0.2
langchain-openai==0.0.2
openai>=1.0.0
chromadb==0.4.18
sentence-transformers==2.2.2
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
tiktoken>=0.5.2,<0.6.0
numpy==1.24.3
pandas==2.0.3
streamlit==1.28.1
beautifulsoup4==4.12.2
markdown==3.5.1
PyPDF2==3.0.1
python-docx==1.1.0 

